backend/pyproject.toml
[project]
name = "reportrx-backend"
version = "0.1.0"
description = "ReportRx FastAPI backend"
requires-python = ">=3.11"
authors = [{name = "ReportRx", email = "dev@reportrx.local"}]
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.7.0",
  "pymupdf>=1.24.0",
  "httpx>=0.27.0",
  "pytesseract>=0.3.10",
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.2",
  "black>=24.4.0",
  "pytest>=8.2.0",
  "httpx>=0.27.0"
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "UP", # pyupgrade
]
ignore = []

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
